UiStore: added notification
