--- webapp/app/appNotification/reducer.js
@@ -1,13 +1,0 @@
-import { exportReducer } from '../../utils/reduxUtils'
-
-import { appNotificationShow, appNotificationHide } from './actions'
-
-import * as NotificationState from './appNotificationState'
-
-const actionHandlers = {
-  [appNotificationShow]: (state, { notification }) => NotificationState.show(notification),
-
-  [appNotificationHide]: NotificationState.hide,
-}
-
-export default exportReducer(actionHandlers)
