+++ webapp/store/ui/notification/index.js
@@ -1,0 +1,5 @@
+import * as NotificationActions from './actions'
+import * as NotificationState from './state'
+import NotificationReducer from './reducer'
+
+export { NotificationActions, NotificationReducer, NotificationState }
