+++ webapp/store/ui/notification/state.js
@@ -1,8 +1,10 @@
 import * as R from 'ramda'
 
-export const stateKey = 'appNotification'
+import * as UiState from '../state'
 
-const getState = R.propOr({}, stateKey)
+export const stateKey = 'notification'
+
+const getState = R.pipe(UiState.getState, R.propOr({}, stateKey))
 
 const keys = {
   messageKey: 'messageKey',
@@ -11,26 +13,29 @@ const keys = {
   visible: 'visible',
 }
 
-export const severity = {
+export const severityType = {
   info: 'info',
   warning: 'warning',
   error: 'error',
 }
 
-export const newNotification = (key, params, severity) => ({
+// ====== CREATE
+export const newNotification = ({ key, params, severity }) => ({
   [keys.messageKey]: key,
   [keys.messageParams]: params,
   [keys.severity]: severity,
 })
 
+// ====== READ
 export const getMessageKey = R.pipe(getState, R.propOr(null, keys.messageKey))
 export const getMessageParams = R.pipe(getState, R.propOr({}, keys.messageParams))
-export const getSeverity = R.pipe(getState, R.propOr(severity.info, keys.severity))
+export const getSeverity = R.pipe(getState, R.propOr(severityType.info, keys.severity))
 export const isVisible = R.pipe(getState, R.propEq(keys.visible, true))
 
-export const show = notification => ({
+export const show = (notification) => ({
   ...notification,
   [keys.visible]: true,
 })
 
+// ====== UPDATE
 export const hide = () => ({})
