+++ webapp/store/ui/reducerParial.js
@@ -1,0 +1,7 @@
+import { combineReducers } from 'redux'
+
+import { NotificationReducer, NotificationState } from './notification'
+
+export default combineReducers({
+  [NotificationState.stateKey]: NotificationReducer,
+})
