+++ webapp/store/ui/state.js
@@ -1,0 +1,5 @@
+import * as R from 'ramda'
+
+export const stateKey = 'ui'
+
+export const getState = R.propOr({}, stateKey)
